+++ src/helpers/crud.js
@@ -1,0 +1,5 @@
+export const retrieve =
+  (service) =>
+  ({ ...config } = {}) =>
+  (id) =>
+    service.retrieve(id, config)