+++ src/services/__tests__/setup/index.js
@@ -8,4 +8,5 @@ export {
   makeShippoOrderService,
   makeShippoTrackService,
   makeShippoService,
+  makeShippoHelper,
 } from "./instances"