+++ src/services/shippo-order.js
@@ -1,8 +1,10 @@
 import { BaseService } from "medusa-interfaces"
-import { MedusaError } from "medusa-core-utils"
+import { shippoHelper } from "../helpers"
 
 class ShippoOrderService extends BaseService {
   #client
+
+  #error
 
   #fulfillmentService
 
@@ -11,6 +13,8 @@ class ShippoOrderService extends BaseService {
   #manager
 
   #shippo
+
+  #helper
 
   #shippoTransactionService
 
@@ -39,6 +43,10 @@ class ShippoOrderService extends BaseService {
     this.#shippoTransactionService = shippoTransactionService
 
     this.#client = this.#shippo.useClient
+
+    this.#helper = (entity) => shippoHelper().helpers[entity]
+
+    this.#error = (entity) => shippoHelper().helpers.error(entity)
   }
 
   /**
@@ -56,7 +64,16 @@ class ShippoOrderService extends BaseService {
    * @return {Promise.<Object>}
    */
   async fetchByFulfillmentId(fulfillmentId) {
-    const shippoOrderId = await this.#getIdFromFulfillment(fulfillmentId)
+    const shippoOrderId = await this.#helper("fulfillment").shippoId(
+      fulfillmentId
+    )
+
+    if (!shippoOrderId) {
+      return this.#error("shippo_order").notFoundFor([
+        "fulfillment",
+        fulfillmentId,
+      ])
+    }
 
     const shippoOrder = await this.fetch(shippoOrderId).then(async (order) => {
       if (order?.transactions?.length) {
@@ -90,8 +107,12 @@ class ShippoOrderService extends BaseService {
    * @return {Promise.<Object>}
    */
   async fetchPackingSlipByFulfillmentId(fulfillmentId) {
-    const shippoOrderId = await this.#getIdFromFulfillment(fulfillmentId)
-    return this.fetchPackingSlip(shippoOrderId)
+    const shippoOrderId = await this.#helper("fulfillment").shippoId(
+      fulfillmentId
+    )
+    return shippoOrderId
+      ? this.fetchPackingSlip(shippoOrderId)
+      : this.#error("shippo_order").notFoundFor(["fulfillment", fulfillmentId])
   }
 
   /**
@@ -114,41 +135,16 @@ class ShippoOrderService extends BaseService {
     })
 
     if (!fulfillment.length) {
-      return Promise.reject(
-        new MedusaError(
-          MedusaError.Types.NOT_FOUND,
-          `Fulfillment for shippo order with id: ${orderId} not found`
-        )
-      )
+      return this.#error("fulfillment").notFoundFor(["shippo order", orderId])
     }
+
     return fulfillment[0]
   }
 
-  async findFulfillmentsBy(type, id) {
-    const fulfillmentRepo = this.#manager.getCustomRepository(
-      this.#fulfillmentRepo
-    )
-
-    const fulfillments = await fulfillmentRepo.find({
-      where: {
-        [type]: id,
-      },
-    })
-    return fulfillments
-  }
-
   async findBy(type, id) {
-    const fulfillments = await this.findFulfillmentsBy(type, id).then(
-      (response) => response.filter((ful) => ful.data.shippo_order_id)
-    )
+    const fulfillments = await this.#retrieveFulfillments(type, id)
 
-    if (fulfillments.length === 0) {
-      return Promise.reject(
-        new MedusaError(MedusaError.Types.NOT_FOUND, `Shippo order not found`)
-      )
-    }
-
-    const orders = await Promise.all(
+    return Promise.all(
       fulfillments.map(async (fulfillment) => {
         const order = await this.fetchByFulfillmentId(fulfillment.id)
         return {
@@ -157,38 +153,23 @@ class ShippoOrderService extends BaseService {
         }
       })
     )
-    return orders
   }
 
   async findPackingSlipBy(type, id) {
-    const fulfillments = await this.findFulfillmentsBy(type, id).then(
-      (response) => response.filter((ful) => ful.data.shippo_order_id)
-    )
+    const fulfillments = await this.#retrieveFulfillments(type, id)
 
-    if (fulfillments.length === 0) {
-      return Promise.reject(
-        new MedusaError(MedusaError.Types.NOT_FOUND, `Shippo order not found`)
-      )
-    }
-
-    const packingSlips = await Promise.all(
+    return Promise.all(
       fulfillments.map(async (fulfillment) => {
         const packingSlip = await this.fetchPackingSlipByFulfillmentId(
           fulfillment.id
         )
-        const {
-          id,
-          data: { shippo_order_id },
-        } = fulfillment
-
         return {
           ...packingSlip,
-          shippo_order_id,
-          fulfillment_id: id,
+          shippo_order_id: fulfillment.data.shippo_order_id,
+          fulfillment_id: fulfillment.id,
         }
       })
     )
-    return packingSlips
   }
 
   async isReplace(id) {
@@ -196,25 +177,12 @@ class ShippoOrderService extends BaseService {
     return order.order_number.includes("replace")
   }
 
-  async #getIdFromFulfillment(fulfillmentOrId) {
-    const fulfillment = fulfillmentOrId?.id
-      ? fulfillmentOrId
-      : await this.#fulfillmentService.retrieve(fulfillmentOrId)
+  async #retrieveFulfillments(type, id) {
+    const fulfillments = await this.#helper("fulfillment").for(type)(id)
 
-    if (!fulfillment.data?.shippo_order_id) {
-      return Promise.reject(
-        new MedusaError(
-          MedusaError.Types.NOT_FOUND,
-          `Shippo order not found for fulfillment with id: ${fulfillment.id}`
-        )
-      )
-    }
-
-    const {
-      data: { shippo_order_id },
-    } = fulfillment
-
-    return shippo_order_id
+    return fulfillments.length > 0
+      ? fulfillments
+      : this.#error("shippo_order").notFoundFor([type, id])
   }
 }