+++ src/utils/index.js
@@ -1,0 +1,4 @@
+export const subStringBefore = (char) => (str) =>
+  str.substring(0, str.indexOf(char))
+
+export const idType = subStringBefore("_")